+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
+## New Features
 
-# Minor Update
+- Implement V3 Payment endpoints (Create Payment Access Token, Get Accepted Payment Methods)
+
+## Minor Update
 
 - Bump guzzlehttp/guzzle from 7.1.1 to 7.2.0